TODO 1: URL routes
Lined out the URL problems. Actually, fixed a lot of problems in not a lot of
time. It just took a little thinking.

Added a BeforeRender hook to make global config visible to templates. Actually,
this is probably a better way to pass data. Credit goes to eevee.

Still TODO:
-   fix formatting
-   fix db schema
-   more data

I might work with formatting tonight.