Adding pages to the site
Woo, the home page isn't full of broken links any longer.

I changed the site intro in home.mako, wrote a description of the project in
about.mako, and added a wufoo contact form in contact.mako (didn't feel like
messing with an e-mail form yet... TODO).

The changes in the other templates are just me changing the title blocks to
defs, which it turns out are kind of easier to work with.